refactor: EventHistoryController 생성
EventDiscountController의 로직 분리
