+++ src/main/java/christmas/controller/EventDiscountController.java
@@ -83,56 +83,4 @@ public class EventDiscountController {
                 order.isMeetCriteria(PresentEventInfo.CHAMPAGNE_EVENT.getMoneyCriteria());
     }
 
-    private String getChristmasEventHistory(VisitDate visitDate, Order order) {
-        int totalDiscount = calculateChristmasDiscount(visitDate, order);
-        String eventHistory = getEventHistoryForm(ReportPhrase.CHRISTMAS_DISCOUNT.getPhrase(), totalDiscount);
-        return getBlankOrHistoryByTotalDiscount(totalDiscount, eventHistory);
-    }
-
-    private String getWeekdayEventHistory(VisitDate visitDate, Order order) {
-        int totalDiscount = calculateWeekdayDiscount(visitDate, order);
-        String eventHistory = getEventHistoryForm(ReportPhrase.WEEKDAY_DISCOUNT.getPhrase(), totalDiscount);
-        return getBlankOrHistoryByTotalDiscount(totalDiscount, eventHistory);
-    }
-
-    private String getWeekendEventHistory(VisitDate visitDate, Order order) {
-        int totalDiscount = calculateWeekendDiscount(visitDate, order);
-        String eventHistory = getEventHistoryForm(ReportPhrase.WEEKEND_DISCOUNT.getPhrase(), totalDiscount);
-        return getBlankOrHistoryByTotalDiscount(totalDiscount, eventHistory);
-    }
-
-    private String getSpecialEventHistory(VisitDate visitDate, Order order) {
-        int totalDiscount = calculateSpecialDiscount(visitDate, order);
-        String eventHistory = getEventHistoryForm(ReportPhrase.SPECIAL_DISCOUNT.getPhrase(), totalDiscount);
-        return getBlankOrHistoryByTotalDiscount(totalDiscount, eventHistory);
-    }
-
-    private String getPresentEventHistory(VisitDate visitDate, Order order) {
-        int totalDiscount = calculatePresentDiscount(visitDate, order);
-        String eventHistory = getEventHistoryForm(ReportPhrase.PRESENT_EVENT.getPhrase(), totalDiscount);
-        return getBlankOrHistoryByTotalDiscount(totalDiscount, eventHistory);
-    }
-
-    private String getEventHistoryForm(String title, int totalDiscount) {
-        return title + ReportPhrase.COLON.getPhrase() +
-                ReportPhrase.MINUS.getPhrase() + Parser.toThousandUnitMoney(totalDiscount) + "\n";
-    }
-
-    private String getBlankOrHistoryByTotalDiscount(int totalDiscount, String eventHistory) {
-        if (totalDiscount == 0) {
-            return "";
-        }
-        return eventHistory;
-    }
-
-    public String getEventHistory(VisitDate visitDate, Order order) {
-        StringBuilder eventHistory = new StringBuilder();
-        eventHistory.append(getChristmasEventHistory(visitDate, order));
-        eventHistory.append(getWeekdayEventHistory(visitDate, order));
-        eventHistory.append(getWeekendEventHistory(visitDate, order));
-        eventHistory.append(getSpecialEventHistory(visitDate, order));
-        eventHistory.append(getPresentEventHistory(visitDate, order));
-
-        return eventHistory.toString();
-    }
 }
